requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-tinymce==3.5.0
django-versatileimagefield==2.2
Pillow==9.4.0
python-decouple==3.6
python-magic==0.4.27
sqlparse==0.4.3
